Remove the remaining shift and null-pointer-arithmetic UB
- MvDK deflate: the chunk decoder now tracks output as offsets, so the
  validator no longer sizes the stream by doing arithmetic on a NULL
  destination pointer; the stored-chunk source check no longer forms a
  pointer up to 2 GiB past the input. The canonical Huffman path
  accumulator is unsigned (corrupt depth counts overflowed an int).
- BRSUB_MODE2/3 shifted file-format enums >= 128 left by 24 as int.
- Signed byte << 24 in lib-bcres.c, lib-mvdk.c and lib-plt0.c.
- DSP-ADPCM decode left-shifted a negative sample; multiply instead.

A UBSan build (shift-base, shift-exponent, pointer-overflow,
signed-integer-overflow) now reports nothing over the fixture corpus plus
mutants; extraction output and tests/regress.sh results are unchanged.

+++ project/src/lib-bcres.c
@@ -20,7 +20,7 @@ static uint32_t read_u32 (bcres_stream_t *stream)
 	if (stream->pos + 4 > stream->size)
 		return 0;
 	uint32_t val = stream->data[stream->pos] | (stream->data[stream->pos + 1] << 8)
-		| (stream->data[stream->pos + 2] << 16) | (stream->data[stream->pos + 3] << 24);
+		| (stream->data[stream->pos + 2] << 16) | ((u32)stream->data[stream->pos + 3] << 24);
 	stream->pos += 4;
 	return val;
 }
