+++ project/src/lib-plt0.c
@@ -19,8 +19,8 @@ enumError LoadPLT0 (Image_t *img, const u8 *data, uint data_size)
 	if (!data || data_size < 0x20 || memcmp (data, "PLT0", 4) != 0)
 		return ERR_INVALID_DATA;
 
-	u32 pal_offset = (data[0x10] << 24) | (data[0x11] << 16) | (data[0x12] << 8) | data[0x13];
-	u32 pform = (data[0x18] << 24) | (data[0x19] << 16) | (data[0x1A] << 8) | data[0x1B];
+	u32 pal_offset = ((u32)data[0x10] << 24) | (data[0x11] << 16) | (data[0x12] << 8) | data[0x13];
+	u32 pform = ((u32)data[0x18] << 24) | (data[0x19] << 16) | (data[0x1A] << 8) | data[0x1B];
 	u16 num_colors = (data[0x1C] << 8) | data[0x1D];
 
 	if (!num_colors || (u64)pal_offset + (u64)num_colors * 2 > data_size)
@@ -125,8 +125,8 @@ bool GetRawPLT0 (
 	if (!data || data_size < 0x20 || memcmp (data, "PLT0", 4) != 0)
 		return false;
 
-	u32 pal_offset = (data[0x10] << 24) | (data[0x11] << 16) | (data[0x12] << 8) | data[0x13];
-	u32 raw_pform = (data[0x18] << 24) | (data[0x19] << 16) | (data[0x1A] << 8) | data[0x1B];
+	u32 pal_offset = ((u32)data[0x10] << 24) | (data[0x11] << 16) | (data[0x12] << 8) | data[0x13];
+	u32 raw_pform = ((u32)data[0x18] << 24) | (data[0x19] << 16) | (data[0x1A] << 8) | data[0x1B];
 	u16 num_colors = (data[0x1C] << 8) | data[0x1D];
 
 	if (!num_colors || (u64)pal_offset + (u64)num_colors * 2 > data_size)
